Revert "Cryo tank SFX & misc machinery code cleanup (#14436)"
This reverts commit 1f47b818632ecaf7592c5af449e27f680f30ac7f.
