+++ code/game/machinery/computer/camera.dm
@@ -218,6 +218,7 @@
 	canSmoothWith = null
 
 	clockwork = TRUE //it'd look weird
+	broken_overlay_emissive = TRUE
 	pass_flags = PASSTABLE
 
 /obj/machinery/computer/security/mining
@@ -286,6 +287,7 @@
 	density = FALSE
 	circuit = null
 	clockwork = TRUE //it'd look very weird
+	broken_overlay_emissive = TRUE
 	light_power = 0
 
 /obj/machinery/computer/security/telescreen/update_icon()
