+++ code/game/machinery/fat_sucker.dm
@@ -19,18 +19,18 @@
 	var/breakout_time = 600
 
 	var/next_fact = 10 //in ticks, so about 20 seconds
-	var/static/list/fat_facts = list(
-		"Fats are triglycerides made up of a combination of different building blocks; glycerol and fatty acids.",
-		"Adults should get a recommended 20-35% of their energy intake from fat.",
-		"Being overweight or obese puts you at an increased risk of chronic diseases, such as cardiovascular diseases, metabolic syndrome, type 2 diabetes and some types of cancers.",
-		"Not all fats are bad. A certain amount of fat is an essential part of a healthy balanced diet. ",
-		"Saturated fat should form no more than 11% of your daily calories.",
-		"Unsaturated fat, that is monounsaturated fats, polyunsaturated fats and omega-3 fatty acids, is found in plant foods and fish.",
+	var/static/list/fat_facts = list(\
+	"Fats are triglycerides made up of a combination of different building blocks; glycerol and fatty acids.", \
+	"Adults should get a recommended 20-35% of their energy intake from fat.", \
+	"Being overweight or obese puts you at an increased risk of chronic diseases, such as cardiovascular diseases, metabolic syndrome, type 2 diabetes and some types of cancers.", \
+	"Not all fats are bad. A certain amount of fat is an essential part of a healthy balanced diet. " , \
+	"Saturated fat should form no more than 11% of your daily calories.", \
+	"Unsaturated fat, that is monounsaturated fats, polyunsaturated fats and omega-3 fatty acids, is found in plant foods and fish." \
 	)
 
 /obj/machinery/fat_sucker/Initialize(mapload)
 	. = ..()
-	soundloop = new(src, FALSE)
+	soundloop = new(src,  FALSE)
 	update_icon()
 
 /obj/machinery/fat_sucker/Destroy()
@@ -50,11 +50,11 @@
 
 /obj/machinery/fat_sucker/examine(mob/user)
 	. = ..()
-	. += span_notice("Alt-Click to toggle the safety hatch.")
-	. += span_notice("Removing [bite_size] nutritional units per operation.")
-	. += span_notice("Requires [nutrient_to_meat] nutritional units per meat slab.")
+	. += "[span_notice("Alt-Click to toggle the safety hatch.")]\n"+\
+			"[span_notice("Removing [bite_size] nutritional units per operation.")]\n"+\
+			span_notice("Requires [nutrient_to_meat] nutritional units per meat slab.")
 
-/obj/machinery/fat_sucker/close_machine(mob/user, density_to_set = TRUE)
+/obj/machinery/fat_sucker/close_machine(mob/user)
 	if(panel_open)
 		to_chat(user, span_warning("You need to close the maintenance hatch first!"))
 		return
