+++ code/game/machinery/porta_turret/portable_turret_cover.dm
@@ -83,6 +83,9 @@ DEFINE_BUFFER_HANDLER(/obj/machinery/porta_turret_cover)
 /obj/machinery/porta_turret_cover/attack_hulk(mob/living/carbon/human/user, does_attack_animation = 0)
 	return parent_turret.attack_hulk(user)
 
+/obj/machinery/porta_turret_cover/can_be_overridden()
+	. = 0
+
 /obj/machinery/porta_turret_cover/should_emag(mob/user)
 	return parent_turret.should_emag(user)
