+++ code/game/machinery/rechargestation.dm
@@ -77,11 +77,11 @@
 	else
 		open_machine()
 
-/obj/machinery/recharge_station/open_machine(drop = TRUE, density_to_set = FALSE)
+/obj/machinery/recharge_station/open_machine()
 	. = ..()
 	update_use_power(IDLE_POWER_USE)
 
-/obj/machinery/recharge_station/close_machine(mob/user, density_to_set = TRUE)
+/obj/machinery/recharge_station/close_machine()
 	. = ..()
 	if(occupant)
 		update_use_power(ACTIVE_POWER_USE) //It always tries to charge, even if it can't.
