+++ code/game/machinery/sleeper.dm
@@ -134,12 +134,12 @@
 		container_resist(user)
 
 //Note: open_machine and close_machine already ui_update()
-/obj/machinery/sleeper/open_machine(drop = TRUE, density_to_set = FALSE)
+/obj/machinery/sleeper/open_machine()
 	if(!state_open && !panel_open)
 		flick("[initial(icon_state)]-anim", src)
 		..()
 
-/obj/machinery/sleeper/close_machine(mob/user, density_to_set = TRUE)
+/obj/machinery/sleeper/close_machine(mob/user)
 	if((isnull(user) || istype(user)) && state_open && !panel_open)
 		flick("[initial(icon_state)]-anim", src)
 		..(user)
