+++ code/game/machinery/washing_machine.dm
@@ -501,7 +501,7 @@ GLOBAL_LIST_INIT(dye_registry, list(
 		new /obj/item/stack/sheet/iron(drop_location(), 2)
 	qdel(src)
 
-/obj/machinery/washing_machine/open_machine(drop = TRUE, density_to_set = FALSE)
+/obj/machinery/washing_machine/open_machine(drop = 1)
 	..()
 	set_density(TRUE) //because machinery/open_machine() sets it to FALSE
 	color_source = null
