+++ code/game/objects/items/implants/implantchair.dm
@@ -149,7 +149,7 @@
 	close_machine(target)
 
 
-/obj/machinery/implantchair/close_machine(mob/living/user, density_to_set = TRUE)
+/obj/machinery/implantchair/close_machine(mob/living/user)
 	if((isnull(user) || istype(user)) && state_open)
 		..(user)
 		if(auto_inject && ready && ready_implants > 0)
