+++ code/modules/antagonists/abductor/machinery/experiment.dm
@@ -22,16 +22,16 @@
 		return
 	close_machine(target)
 
-/obj/machinery/abductor/experiment/open_machine(drop = TRUE, density_to_set = FALSE)
+/obj/machinery/abductor/experiment/open_machine()
 	if(!state_open && !panel_open)
 		..()
 
-/obj/machinery/abductor/experiment/close_machine(mob/target, density_to_set = TRUE)
+/obj/machinery/abductor/experiment/close_machine(mob/target)
 	for(var/A in loc)
 		if(isabductor(A))
 			return
 	if(state_open && !panel_open)
-		..()
+		..(target)
 
 /obj/machinery/abductor/experiment/relaymove(mob/living/user, direction)
 	if(user.stat != CONSCIOUS)
