+++ code/modules/mob/living/carbon/life.dm
@@ -67,7 +67,7 @@
 		else
 			SEND_SIGNAL(src, COMSIG_CLEAR_MOOD_EVENT, "suffocation")
 	else
-		if(isobj(loc))
+		if(istype(loc, /obj/))
 			var/obj/location_as_object = loc
 			location_as_object.handle_internal_lifeform(src,0)
 
@@ -95,7 +95,7 @@
 		losebreath--
 		if(prob(10))
 			emote("gasp")
-		if(isobj(loc))
+		if(istype(loc, /obj/))
 			var/obj/loc_as_obj = loc
 			loc_as_obj.handle_internal_lifeform(src,0)
 	else
