+++ code/modules/research/nanites/nanite_chamber.dm
@@ -180,14 +180,14 @@
 			span_notice("You successfully break out of [src]!"))
 		open_machine()
 
-/obj/machinery/nanite_chamber/close_machine(mob/living/carbon/user, density_to_set = TRUE)
+/obj/machinery/nanite_chamber/close_machine(mob/living/carbon/user)
 	if(!state_open)
 		return FALSE
 
 	..(user)
 	return TRUE
 
-/obj/machinery/nanite_chamber/open_machine(drop = TRUE, density_to_set = FALSE)
+/obj/machinery/nanite_chamber/open_machine()
 	if(state_open)
 		return FALSE
